refactor: update mappers names
