+++ src/components/picture/Picture.astro
@@ -1,7 +1,7 @@
 ---
 import { DEFAULT_ALT_TEXT } from "../../constants/defaults";
 import type { CmsImage } from "../../lib/content/models/image";
-import FormattedDate from "../FormattedDate.astro";
+import FormattedDate from "../formatted-date/FormattedDate.astro";
 import { buildCmsPictureAttributes } from "./Picture.helpers";
 
 interface Props {
