+++ src/layouts/BlogPost.astro
@@ -1,7 +1,7 @@
 ---
 import { Image } from "astro:assets";
-import FormattedDate from "../components/FormattedDate.astro";
 import Footer from "../components/footer/Footer.astro";
+import FormattedDate from "../components/formatted-date/FormattedDate.astro";
 import Header from "../components/header/Header.astro";
 import MetaData from "../components/meta-data/MetaData.astro";
 import type {
