+++ src/lib/content/mappers/site-config.ts
@@ -3,9 +3,9 @@ import {
 	DEFAULT_SITE_TITLE,
 } from "../../../constants/defaults";
 import type { SiteConfig } from "../models";
-import { normalizeRequiredString } from "./helpers";
 import { mapCmsImage } from "./image";
 import { mapLinks } from "./link";
+import { normalizeRequiredString } from "./normalizers";
 import { mapRichTextBlock } from "./rich-text-block";
 
 export function mapSiteConfig(
