+++ src/lib/content/mappers/structured-data.ts
@@ -1,9 +1,5 @@
 import type { JsonValue } from "../models/json";
 
-/**
- * Produces safe inner HTML for `<script type="application/ld+json">`.
- * Returns `null` when there is nothing valid to emit.
- */
 export function mapStructuredDataToJsonLdHtml(
 	structuredData: JsonValue | undefined,
 ): string | null {
