+++ src/pages/about.astro
@@ -1,5 +1,5 @@
 ---
-import Layout from "../layouts/SinglePageWithDynamicContent.astro";
+import Layout from "../layouts/DynamicContentPage.astro";
 import { createContentProvider } from "../lib/content/provider";
 
 const provider = createContentProvider();
